Explicitly set order state to complete + call finalize, rather than jumping through all the states
Previously, the checkout controller decorator here would go through each of the order states and call @order.update! for each state, as well as the order callback.

This is unecessary for PayPal transactions.

All that needs to happen here is that the state is put into the complete state and the order is finalized.

Fixes #48.

Finally.